requirements.txt
feedparser
google-genai
colorama
rich
openai
requests
python-dotenv
